==> src/2 - Pre-Processed Data Visualization.ipynb <==
# 2 - Pre-Processed Data Visualization

Visualization of the raw data corrected in the previous step.

## Workspace

workspace = {
    "settings": {
        "folder": "E:\\DataSets\\100 Hz"
    },
    "datasets": {
        "saveiro-1": {
            "folder": "PVS 1"
        },
        "saveiro-2": {
            "folder": "PVS 2"
        },
        "saveiro-3": {
            "folder": "PVS 3"
        },
        "bravo-1": {
            "folder": "PVS 4"
        },
        "bravo-2": {
            "folder": "PVS 5"
        },
        "bravo-3": {
            "folder": "PVS 6"
        },
        "palio-1": {
            "folder": "PVS 7"
        },
        "palio-2": {
            "folder": "PVS 8"
        },
        "palio-3": {
            "folder": "PVS 9"
        }
    }
}

## Importing Packages

import os
import folium
import pandas as pd
import matplotlib.pyplot as plt
# %matplotlib inline
%matplotlib notebook
pd.set_option("float_format", '{:0.10f}'.format)
pd.set_option('display.max_columns', 30)

## Utility Functions

# Function to plot the data
def plotData(data, typeAxis, title, raw=False, mean=False, std=False, window=5):
    
    plt.figure(figsize=(16,6)) 
    plt.title(title)
    
    below_suspension = data[typeAxis + '_below_suspension']
    above_suspension = data[typeAxis + '_above_suspension']
    dashboard = data[typeAxis + '_dashboard']
    speed = data['speed']
    
    if raw:
        below_suspension.plot()
        above_suspension.plot()
        dashboard.plot()
        speed.plot()
    
    if mean:
        below_suspension.rolling(window).mean().plot()
        above_suspension.rolling(window).mean().plot()
        dashboard.rolling(window).mean().plot()
        speed.plot()
        
    if std:
        below_suspension.rolling(window).std().plot()
        above_suspension.rolling(window).std().plot()
        dashboard.rolling(window).std().plot()
        speed.plot()
    
    plt.legend(loc='upper center')

## Data Import

# Dataset selection
dataset = workspace['datasets']['saveiro-1']
folder = os.path.join(workspace['settings']['folder'], dataset['folder'])
gps = pd.read_csv(os.path.join(folder, 'dataset_gps.csv'), float_precision="high")
data_Left = pd.read_csv(os.path.join(folder, 'dataset_gps_mpu_left.csv'), float_precision="high")
data_Right = pd.read_csv(os.path.join(folder, 'dataset_gps_mpu_right.csv'), float_precision="high")

## Data Visualization

gps.describe()

data_Left.describe()

data_Right.describe()

gps.head()

data_Left.head()

data_Right.head()

## Acceleration

<h6>Acc X Left</h6>

plotData(data_Left, "acc_x", "Acc X - Left - Raw Data", raw=True)

plotData(data_Left, "acc_x", "Acc X - Left - Moving average", mean=True)

plotData(data_Left, "acc_x", "Acc X - Left - Moving Standard Deviation", std=True)

<h6>Acc X Right</h6>

plotData(data_Right, "acc_x", "Acc X - Right - Raw Data", raw=True)

plotData(data_Right, "acc_x", "Acc X - Right - Moving average", mean=True)

plotData(data_Right, "acc_x", "Acc X - Right - Moving Standard Deviation", std=True)

<h6>Acc Y Left</h6>

plotData(data_Left, "acc_y", "Acc Y - Left - Raw Data", raw=True)

plotData(data_Left, "acc_y", "Acc Y - Left - Moving average", mean=True)

plotData(data_Left, "acc_y", "Acc Y - Left - Moving Standard Deviation", std=True)

<h6>Acc Y Right</h6>

plotData(data_Right, "acc_y", "Acc Y - Right - Raw Data", raw=True)

plotData(data_Right, "acc_y", "Acc Y - Right - Moving average", mean=True)

plotData(data_Right, "acc_y", "Acc Y - Right - Moving Standard Deviation", std=True)

<h6>Acc Z Left</h6>

plotData(data_Left, "acc_z", "Acc z - Left - Raw Data", raw=True)

plotData(data_Left, "acc_z", "Acc Z - Left - Moving average", mean=True)

plotData(data_Left, "acc_z", "Acc Z - Left - Moving Standard Deviation", std=True)

<h6>Acc Z Right</h6>

plotData(data_Right, "acc_z", "Acc Z - Right - Raw Data", raw=True)

plotData(data_Right, "acc_z", "Acc Z - Right - Moving average", mean=True)

plotData(data_Right, "acc_z", "Acc Z - Right - Moving Standard Deviation", std=True)

## Gyroscope

<h6>Gyro X Left</h6>

plotData(data_Left, "gyro_x", "Gyro X - Left - Raw Data", raw=True)

plotData(data_Left, "gyro_x", "Gyro X - Left - Moving average", mean=True)

plotData(data_Left, "gyro_x", "Gyro X - Left - Moving Standard Deviation", std=True)

<h6>Gyro X Right</h6>

plotData(data_Right, "gyro_x", "Gyro X - Right - Raw Data", raw=True)

plotData(data_Right, "gyro_x", "Gyro X - Right - Moving average", mean=True)

plotData(data_Right, "gyro_x", "Gyro X - Right - Moving Standard Deviation", std=True)

<h6>Gyro Y Left</h6>

plotData(data_Left, "gyro_y", "Gyro Y - Left - Raw Data", raw=True)

plotData(data_Left, "gyro_y", "Gyro Y - Left - Moving average", mean=True)

plotData(data_Left, "gyro_y", "Gyro Y - Left - Moving Standard Deviation", std=True)

<h6>Gyro Y Right</h6>

plotData(data_Right, "gyro_y", "Gyro Y - Right - Raw Data", raw=True)

plotData(data_Right, "gyro_y", "Gyro Y - Right - Moving average", mean=True)

plotData(data_Right, "gyro_y", "Gyro Y - Right - Moving Standard Deviation", std=True)

<h6>Gyro Z Left</h6>

plotData(data_Left, "gyro_z", "Gyro Z - Left - Raw Data", raw=True)

plotData(data_Left, "gyro_z", "Gyro Z - Left - Moving average", mean=True)

plotData(data_Left, "gyro_z", "Gyro Z - Left - Moving Standard Deviation", std=True)

<h6>Gyro Z Right</h6>

plotData(data_Right, "gyro_z", "Gyro Z - Right - Raw Data", raw=True)

plotData(data_Right, "gyro_z", "Gyro Z - Right - Moving average", mean=True)

plotData(data_Right, "gyro_z", "Gyro Z - Right - Moving Standard Deviation", std=True)

## GPS

speedMS = gps['speed_meters_per_second']
speedKMH = gps['speed_meters_per_second'] * 3.6
speedEstimated = gps['distance_meters'] / gps['elapsed_time_seconds']

plt.figure(figsize=(16,8)) 
plt.title('Speed in m/s and km/h')
speedEstimated.plot(label='m/s - Estimated', color='red')
speedMS.plot(label='m/s')
speedKMH.plot(label='km/h')
plt.legend(loc='upper center')

locations = gps[['latitude', 'longitude']]
coordinates = [tuple(x) for x in locations.to_numpy()]

focolat = (gps['latitude'].min() + gps['latitude'].max()) / 2
focolon = (gps['longitude'].min() + gps['longitude'].max()) / 2
maps = folium.Map(location=[focolat, focolon], zoom_start=14)

folium.PolyLine(coordinates, color="blue", weight=2.5, opacity=1).add_to(maps)

maps
