==> src/2 - Processed Data Visualization.ipynb <==
<h1>Importação e Visualização de Dados Brutos Corrigidos</h1>

Visualização dos dados brutos corrigidos na etapa anterior.

<h4>Importação de Pacotes</h4>

import os
import folium
import pandas as pd
import matplotlib.pyplot as plt
# %matplotlib inline
%matplotlib notebook
pd.set_option("float_format", '{:0.10f}'.format)
pd.set_option('display.max_columns', 30)
%run "2 - Processed Data Visualization.py"

<h4>Funções Utilitárias</h4>

# Plot de Dados
def plotData(data, typeAxis, title, raw=False, mean=False, std=False, window=5):
    
    plt.figure(figsize=(16,6)) 
    plt.title(title)
    
    below_suspension = data[typeAxis + '_below_suspension']
    above_suspension = data[typeAxis + '_above_suspension']
    dashboard = data[typeAxis + '_dashboard']
    speed = data['speed']
    
    if raw:
        below_suspension.plot()
        above_suspension.plot()
        dashboard.plot()
        speed.plot()
    
    if mean:
        below_suspension.rolling(window).mean().plot()
        above_suspension.rolling(window).mean().plot()
        dashboard.rolling(window).mean().plot()
        speed.plot()
        
    if std:
        below_suspension.rolling(window).std().plot()
        above_suspension.rolling(window).std().plot()
        dashboard.rolling(window).std().plot()
        speed.plot()
    
    plt.legend(loc='upper center')

<h4>Importação dos Dados</h4>

# Seleção do Dataset
dataset = workspace['datasets']['saveiro-1']
folder = os.path.join(workspace['settings']['folder'], dataset['folder'])
gps = pd.read_csv(os.path.join(folder, 'dataset_gps.csv'), float_precision="high")
data_esquerda = pd.read_csv(os.path.join(folder, 'dataset_gps_mpu_left.csv'), float_precision="high")
data_direita = pd.read_csv(os.path.join(folder, 'dataset_gps_mpu_right.csv'), float_precision="high")

<h4>Visualização de Dados</h4>

gps.describe()

data_esquerda.describe()

data_direita.describe()

gps.head()

data_esquerda.head()

data_direita.head()

<h2>Acceleration</h2>

<h6>Acc X Esquerda</h6>

plotData(data_esquerda, "acc_x", "Acc X - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "acc_x", "Acc X - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "acc_x", "Acc X - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Acc X Direita</h6>

plotData(data_direita, "acc_x", "Acc X - Direita - Dados Brutos", raw=True)

plotData(data_direita, "acc_x", "Acc X - Direita - Média Móvel", mean=True)

plotData(data_direita, "acc_x", "Acc X - Direita - Desvio Padrão Móvel", std=True)

<h6>Acc Y Esquerda</h6>

plotData(data_esquerda, "acc_y", "Acc Y - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "acc_y", "Acc Y - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "acc_y", "Acc Y - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Acc Y Direita</h6>

plotData(data_direita, "acc_y", "Acc Y - Direita - Dados Brutos", raw=True)

plotData(data_direita, "acc_y", "Acc Y - Direita - Média Móvel", mean=True)

plotData(data_direita, "acc_y", "Acc Y - Direita - Desvio Padrão Móvel", std=True)

<h6>Acc Z Esquerda</h6>

plotData(data_esquerda, "acc_z", "Acc z - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "acc_z", "Acc Z - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "acc_z", "Acc Z - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Acc Z Direita</h6>

plotData(data_direita, "acc_z", "Acc Z - Direita - Dados Brutos", raw=True)

plotData(data_direita, "acc_z", "Acc Z - Direita - Média Móvel", mean=True)

plotData(data_direita, "acc_z", "Acc Z - Direita - Desvio Padrão Móvel", std=True)

<h2>Gyroscope</h2>

<h6>Gyro X Esquerda</h6>

plotData(data_esquerda, "gyro_x", "Gyro X - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "gyro_x", "Gyro X - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "gyro_x", "Gyro X - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Gyro X Direita</h6>

plotData(data_direita, "gyro_x", "Gyro X - Direita - Dados Brutos", raw=True)

plotData(data_direita, "gyro_x", "Gyro X - Direita - Média Móvel", mean=True)

plotData(data_direita, "gyro_x", "Gyro X - Direita - Desvio Padrão Móvel", std=True)

<h6>Gyro Y Esquerda</h6>

plotData(data_esquerda, "gyro_y", "Gyro Y - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "gyro_y", "Gyro Y - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "gyro_y", "Gyro Y - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Gyro Y Direita</h6>

plotData(data_direita, "gyro_y", "Gyro Y - Direita - Dados Brutos", raw=True)

plotData(data_direita, "gyro_y", "Gyro Y - Direita - Média Móvel", mean=True)

plotData(data_direita, "gyro_y", "Gyro Y - Direita - Desvio Padrão Móvel", std=True)

<h6>Gyro Z Esquerda</h6>

plotData(data_esquerda, "gyro_z", "Gyro Z - Esquerda - Dados Brutos", raw=True)

plotData(data_esquerda, "gyro_z", "Gyro Z - Esquerda - Média Móvel", mean=True)

plotData(data_esquerda, "gyro_z", "Gyro Z - Esquerda - Desvio Padrão Móvel", std=True)

<h6>Gyro Z Direita</h6>

plotData(data_direita, "gyro_z", "Gyro Z - Direita - Dados Brutos", raw=True)

plotData(data_direita, "gyro_z", "Gyro Z - Direita - Média Móvel", mean=True)

plotData(data_direita, "gyro_z", "Gyro Z - Direita - Desvio Padrão Móvel", std=True)

<h2>GPS</h2>

speedMS = gps['speed_meters_per_second']
speedKMH = gps['speed_meters_per_second'] * 3.6
speedEstimated = gps['distance_meters'] / gps['elapsed_time_seconds']

plt.figure(figsize=(16,8)) 
plt.title('Speed in m/s and km/h')
speedEstimated.plot(label='m/s - Estimated', color='red')
speedMS.plot(label='m/s')
speedKMH.plot(label='km/h')
plt.legend(loc='upper center')

locations = gps[['latitude', 'longitude']]
coordinates = [tuple(x) for x in locations.to_numpy()]

focolat = (gps['latitude'].min() + gps['latitude'].max()) / 2
focolon = (gps['longitude'].min() + gps['longitude'].max()) / 2
maps = folium.Map(location=[focolat, focolon], zoom_start=14)

folium.PolyLine(coordinates, color="blue", weight=2.5, opacity=1).add_to(maps)

maps
