==> src/2 - Processed Data Visualization.py <==
workspace = {
    "settings": {
        "folder": "E:\\Dados Processados\\100 Hz"
    },
    "datasets": {
        "saveiro-1": {
            "folder": "PVS 1"
        },
        "saveiro-2": {
            "folder": "PVS 2"
        },
        "saveiro-3": {
            "folder": "PVS 3"
        },
        "bravo-1": {
            "folder": "PVS 4"
        },
        "bravo-2": {
            "folder": "PVS 5"
        },
        "bravo-3": {
            "folder": "PVS 6"
        },
        "palio-1": {
            "folder": "PVS 7"
        },
        "palio-2": {
            "folder": "PVS 8"
        },
        "palio-3": {
            "folder": "PVS 9"
        }
    }
}
